627 Add dockerized script to fetch & transform 2018 ACS data
This commit introduces a command to download 2018 ACS data,
transformed into the schema of the nyc_acs table, as a CSV.

It uses a Python script to do so, and uses Docker to support
running the Python script in a controlled environment.

The Python script was a rework of the script previously created to fetch
2017 ACS data: https://github.com/NYCPlanning/db-ceqr-acs/blob/master/2017/ACS%20Merge.ipynb

Note that future commits will introduce commands and scripts
to load the CSV data into the CEQR database.